third_party/blink/web_tests/external/wpt/scroll-animations/css/progress-based-animation-animation-longhand-properties.tentative-expected.txt

This is a testharness.js-based test.
PASS animation-duration
FAIL animation-duration: 0s assert_equals: expected "100px" but got "none"
PASS animation-iteration-count
FAIL animation-iteration-count: 0 assert_equals: expected "0px" but got "none"
FAIL animation-iteration-count: infinite assert_equals: expected "100px" but got "none"
PASS animation-direction: normal
PASS animation-direction: reverse
PASS animation-direction: alternate
PASS animation-direction: alternate-reverse
PASS animation-delay with a positive value
PASS animation-delay with a negative value
PASS animation-fill-mode
Harness: the test ran to completion.
